+++ frontend/src/pages/ManagerLogin/ManagerLogin.tsx
@@ -2,8 +2,8 @@ import { Link } from 'react-router-dom';
 import Header from 'components/Header/Header';
 import Layout from 'components/Layout/Layout';
 import PATH from 'constants/path';
-import LoginForm from './LoginForm';
 import * as Styled from './ManagerLogin.styles';
+import LoginForm from './units/LoginForm';
 
 const ManagerLogin = (): JSX.Element => {
   return (
